+++ packages/remote-sql-libsql/src/index.ts
@@ -1,5 +1,5 @@
 import { RemoteSQL, SQLPreparedStatement, SQLResult, SQLStatement } from 'remote-sql';
-import type { Client } from '@libsql/core/api';
+import type { Client, ResultSet } from '@libsql/core/api';
 
 export class LibSQLPreparedStatement implements SQLPreparedStatement {
 	public readonly args: any[];
@@ -13,12 +13,16 @@ export class LibSQLPreparedStatement implements SQLPreparedStatement {
 	bind(...values: any[]): SQLPreparedStatement {
 		return new LibSQLPreparedStatement(this.client, this.sql, this.args.concat(values));
 	}
-	all<T = any>(): Promise<SQLResult<T>> {
+	async all<T = any>(): Promise<SQLResult<T>> {
+		let resultSet: ResultSet;
 		if (this.args.length > 0) {
-			return this.client.execute({ sql: this.sql, args: this.args }) as unknown as Promise<SQLResult<T>>;
+			resultSet = await this.client.execute({ sql: this.sql, args: this.args });
 		} else {
-			return this.client.execute(this.sql) as unknown as Promise<SQLResult<T>>;
+			resultSet = await this.client.execute(this.sql);
 		}
+		return {
+			results: resultSet.rows,
+		} as SQLResult<T>;
 	}
 }
 
@@ -28,11 +32,12 @@ export class RemoteLibSQL implements RemoteSQL {
 		return new LibSQLPreparedStatement(this.client, sql);
 	}
 	async batch<T = any>(list: SQLPreparedStatement[]): Promise<SQLResult<T>[]> {
-		return this.client.batch(
+		const response = await this.client.batch(
 			list.map((v) => {
 				const p = v as LibSQLPreparedStatement;
 				return p.args.length > 0 ? { sql: p.sql, args: p.args } : p.sql;
 			}),
-		) as unknown as Promise<SQLResult<T>[]>;
+		);
+		return response.map((res) => ({ results: res.rows })) as SQLResult<T>[];
 	}
 }