+++ packages/remote-sql-libsql/CHANGELOG.md
@@ -1,5 +1,13 @@
 # remote-sql-libsql
 
+## 0.0.2
+
+### Patch Changes
+
+- use cloudflare api: {results}
+- Updated dependencies
+  - remote-sql@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes