+++ packages/remote-sql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # remote-sql
 
+## 0.0.2
+
+### Patch Changes
+
+- use cloudflare api: {results}
+
 ## 0.0.1
 
 ### Patch Changes